+++ tencentcloud/connectivity/client.go
@@ -20,6 +20,7 @@ import (
 	"github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/common"
 	"github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/common/profile"
 	cvm "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/cvm/v20170312"
+	cynosdb "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/cynosdb/v20190107"
 	dayu "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/dayu/v20180709"
 	dc "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/dc/v20180410"
 	es "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/es/v20180416"
@@ -72,6 +73,7 @@ type TencentCloudClient struct {
 	postgreConn   *postgre.Client
 	ckafkaConn    *ckafka.Client
 	auditConn     *audit.Client
+	cynosConn     *cynosdb.Client
 }
 
 // NewClientProfile returns a new ClientProfile
@@ -457,4 +459,17 @@ func (me *TencentCloudClient) UseAuditClient() *audit.Client {
 	me.auditConn.WithHttpTransport(&LogRoundTripper{})
 
 	return me.auditConn
+}
+
+// UseCynosdbClient returns cynosdb client for service
+func (me *TencentCloudClient) UseCynosdbClient() *cynosdb.Client {
+	if me.cynosConn != nil {
+		return me.cynosConn
+	}
+
+	cpf := me.NewClientProfile(300)
+	me.cynosConn, _ = cynosdb.NewClient(me.Credential, me.Region, cpf)
+	me.cynosConn.WithHttpTransport(&LogRoundTripper{})
+
+	return me.cynosConn
 }